+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Ability to configure the reaper internal, by [@msaizar](https://github.com/msaizar).
+- Ability to configure the install check deadline, by [@bradrydzewski](https://github.com/bradrydzewski).
+- Ability to configure the install check interval, by [@bradrydzewski](https://github.com/bradrydzewski).
 
 ## [1.3.0] - 2019-09-11
 ### Added